pyproject.toml
[project]
name = "movierec"
version = "0.1.0"
description = "Movie Recommendation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lightning>=2.6.0",
    "pandas>=2.3.3",
    "scipy>=1.15.3",
    "pyyaml>=6.0",
    "torchvision>=0.24.1",
    "datasets>=4.4.1",
    "torchinfo>=1.8.0",
    "matplotlib>=3.10.8",
    "tensorboard>=2.20.0",
    "hydra-core>=1.3.0",
    "seaborn>=0.13.2",
    "streamlit>=1.52.2",
    "plotly>=6.5.0",
]
